Refactor some component utility functions into the associated service.ts file
