+++ projects/ziti-console-lib/src/lib/features/projectable-forms/edge-router/edge-router-form.component.html
@@ -96,39 +96,35 @@
                 >
                     <lib-qr-code
                             [identity]="formData"
-                            [jwt]="jwt"
-                            [token]="token"
-                            [expiration]="enrollmentExpiration"
+                            [jwt]="formData.jwt"
+                            [token]="formData.token"
+                            [expiration]="formData.enrollmentExpiresAt"
                             [type]="'router'"
                     ></lib-qr-code>
                 </lib-form-field-container>
                 <lib-form-field-container
                         *ngIf="formData.id"
                         [title]="'Associated Services'"
-                        [count]="associatedServices.length"
+                        [count]="svc.associatedServices.length"
                         [class]="'hide-overflow'"
                         [helpText]="'Preview list of all services that this Edge Router is associated with.'"
                 >
                     <lib-preview-list
                         [clickable]="false"
                         [isLoading]="servicesLoading"
-                        [allNames]="associatedServiceNames"
+                        [allNames]="svc.associatedServiceNames"
                     ></lib-preview-list>
-                    <div class="svc-test-result" [ngClass]="{'test-result-open': testResultOpen}">
-                        <div class="close icon-close" (click)="closeTestResult()"></div>
-                        <div class="result-container" [innerHTML]="testResult" ></div>
-                    </div>
                 </lib-form-field-container>
                 <lib-form-field-container
                         *ngIf="formData.id"
                         [title]="'Associated Identities'"
-                        [count]="associatedIdentities.length"
+                        [count]="svc.associatedIdentities.length"
                         [helpText]="'Preview list of all Identities that this Edge Router is associated with.'"
                 >
                     <lib-preview-list
                         [clickable]="false"
                         [isLoading]="identitiesLoading"
-                        [allNames]="associatedIdentityNames"
+                        [allNames]="svc.associatedIdentityNames"
                     ></lib-preview-list>
                 </lib-form-field-container>
                 <lib-form-field-container
@@ -137,7 +133,7 @@
                 >
                     <div class="form-row">
                         <input class="form-field-input" [value]="apiCallURL"/>
-                        <div class="icon-copy copy" (click)="copyToClipboard(apiCallURL)"></div>
+                        <div class="icon-copy copy" (click)="svc.copyToClipboard(apiCallURL)"></div>
                     </div>
                     <lib-json-view *ngIf="formData" [(data)]="apiData" [readOnly]="true" [showCopy]="true"></lib-json-view>
                 </lib-form-field-container>
